pyproject.toml
[project]
name = "dataclean"
version = "0.1.0"
description = "A Streamlit-based data cleaning and analysis tool"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "OfficialDiLi" }
]

dependencies = [
    "streamlit>=1.41.1",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "plotly>=6.2.0",
    "missingno>=0.5.2",
    "ydata-profiling>=4.6.4",
    "streamlit-pandas-profiling>=0.1.0",
    "scikit-learn>=1.4.2",
    "matplotlib>=3.10.0",
    "bcrypt>=4.1.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
